Added feedback to only the notifications page
