+++ prime-angular-frontend/src/app/modules/enrolment/pages/notification-confirmation/notification-confirmation.component.scss
@@ -1,0 +1,5 @@
+.feedback {
+  position: fixed;
+  right: 2rem;
+  bottom: 1rem;
+}
