+++ prime-angular-frontend/src/app/modules/enrolment/shared/components/enrollee-page/enrollee-page.component.html
@@ -2,10 +2,3 @@
           [mode]="mode">
   <ng-content></ng-content>
 </app-page>
-<button class="feedback"
-        mat-raised-button
-        color="primary"
-        (click)="feedback()">
-  <mat-icon class="mr-2">message</mat-icon>
-  Feedback
-</button>
