--- prime-angular-frontend/src/app/modules/enrolment/shared/components/enrollee-page/enrollee-page.component.scss
@@ -1,5 +1,0 @@
-.feedback {
-  position: fixed;
-  right: 2rem;
-  bottom: 1rem;
-}
