+++ prime-angular-frontend/src/app/shared/services/feedback-resource.service.ts
@@ -1,10 +1,13 @@
 import { Injectable } from '@angular/core';
+
+import { Observable } from 'rxjs';
+import { tap, map, catchError } from 'rxjs/operators';
+
+import { ApiHttpResponse } from '@core/models/api-http-response.model';
+import { ToastService } from '@core/services/toast.service';
 import { LoggerService } from '@core/services/logger.service';
 import { ApiResource } from '@core/resources/api-resource.service';
-import { Observable } from 'rxjs';
 import { Feedback } from '@shared/components/dialogs/content/feedback/feedback.component';
-import { tap, map } from 'rxjs/operators';
-import { ApiHttpResponse } from '@core/models/api-http-response.model';
 
 @Injectable({
   providedIn: 'root'
@@ -13,6 +16,7 @@ export class FeedbackResourceService {
 
   constructor(
     private apiResource: ApiResource,
+    private toastService: ToastService,
     private logger: LoggerService
   ) { }
 
@@ -20,7 +24,12 @@ export class FeedbackResourceService {
     return this.apiResource.post(`feedback`, payload)
       .pipe(
         map((response: ApiHttpResponse<Feedback>) => response.result),
-        tap((feedbackReturn: Feedback) => this.logger.info('FEEDBACK', feedbackReturn))
+        tap(() => this.toastService.openSuccessToast('Feedback has been submitted')),
+        catchError((error: any) => {
+          this.toastService.openErrorToast('Feedback could not be submitted');
+          this.logger.error('[Enrolment] FeedbackResourceService::createFeedback error has occurred: ', error);
+          throw error;
+        })
       );
   }
 }
